Don't report probe status to ConnectivityService.
Late in the Q release cycle, NetworkMonitor started sending
probe results as well as evaluation results to
ConnectivityService. This resulted in passing information to
NetworkAgents at the wrong time. For example, as soon as it
connected to a network, it would report INVALID_NETWORK to
the NetworkAgents, causing Wi-Fi to disable autojoin and
telephony to detect a data stall and initiate recovery.

Unfortunately there is not enough information to correctly
suppress these messages in ConnectivityService. So, just
stop sending them. ConnectivityService doesn't use them
anyway.

Bug: 134446021
Test: atest NetworkStackTests
Change-Id: I9e3d9b9057fd017e202d056246ca2711f73d28c7