Add build target for NetworkStack on API 29
NetworkStack will need at least two different build targets: one for the
in-progress SDK for testing before release, and one for the latest
released SDK for releasing to production.

The difference in available APIs is abstracted from the core code by
using shims, with one shim for the released SDK that uses fallback code,
and one shim for the in-progress SDK that calls the actual new API.
The shim included in the built code depends on the build target.

AOSP does not yet have prebuilts for API 29, so the build rule still
references system_current for now.

Test: atest NetworkStackTests
Test: flashed, WiFi working
Test: aapt dump xmltree NetworkStack.apk AndroidManifest.xml: no change
Test: m NetworkStackApiStable (on a branch that has the prebuilt SDK)
Change-Id: I9cf8e712b56a8aad8656eb5095ac8a32f4a05037