Move NetworkStackClient lib to NetworkStack folder
The AIDLs and ipmemorystore-client were already statically linked
libraries, so their location in the source tree does not change anything
(they were already built into the NetworkStack module and the
framework). It makes more sense to have them in the NetworkStack module
folder however, as changes to these files will be released with the same
cadence as the module, not the framework.

This CL creates a new directory and well-defined static library for the
AIDL interface with the NetworkStack, and Java classes used as the
interface

Bug: 139106271
Test: built, flashed, WiFi working
Test: atest FrameworksNetTests FrameworksWifiTests NetworkStackTests
Change-Id: I821b476411287b5cbdf09c82e07bd9ed4889f60d