Add tests for stable API NetworkStack to presubmit
The API stable (system_29 at the moment) version of the NetworkStack
needs to be covered by unit tests in presubmit.

At the moment any difference between NetworkStackNext and
NetworkStack is dead code, so NetworkStackNextTests and
NetworkStackTests are identical.
This change renames NetworkStackTests to NetworkStackNextTests (because
NetworkStackTests historically used the "current", non-stable-API
version of the NetworkStack library), and adds tests for the stable API
version of the code as NetworkStackTests.

Test: atest NetworkStackTests NetworkStackApiStableTests
Change-Id: Iff9a94a7398791adbe856008bc4f9b6c131c3de7