Support dropping RDNSS options with a low lifetime.
Some routers have been known to use RDNSS lifetimes of 10
seconds(!). This causes APF filters to be set to very low
lifetimes, which substantially impacts battery life.

There are two parts to this:
1. Ignore low RDNSS option lifetimes for the purpose of
   calculating APF filter lifetimes.
2. Do not add DNS servers to the repository if their lifetimes
   are too low.

If we do #1 without #2, the servers will expire because APF will
drop the RAs that refresh them, potentially causing outages or
disconnections.

The behaviour is enabled by default starting from R and can be
enabled on all builds using a flag.

Bug: 66928272
Test: New unit test passes
Change-Id: Ib2e2555026da3e81ea3d50767a30092413b4e4f5