Make the APF tests a bit more realistic.
Currently the APF tests include empty options, and manually
assemble the packets. Fix this by using helper methods.

Also increase the memory size of the APF interpreter used by the
test. This is because future CLs in this series add enough RAs to
hit the limit. Also add a log line to the interpreter when the
filter hits the maximum size.

Bug: 66928272
Test: test-only change
Change-Id: Ieed58e278b8a9e7d7f1bdac44aadf56d658df50f