Fix RAs with different retansmission timer would be dropped by apf
When firmware receives RAs with different retransmission
timer, it is expected the RAs should be accepted by apf
filter. However, they are currently dropped since missing
fields which should be added into match section. It causes
to apf filter treats those RAs as the same and then drops.

This change adds the remaining fields to match section to
compare reachable time and retransmission timer with incoming
RAs.

Also, add test to check that RIOs differing only in the
first 4 bytes are different should be passed.

Bug: 143186590
Test: sent RAs with different rtt and check RAs are accepted
Change-Id: I7e2de29740f96b212634b5aeffe709d57afafc68