Use NetworkStack as mainline target w/ stable API
Instead of NetworkStack building against system_current and
NetworkStackApiStable building against system_29, have NetworkStack be
the target built against system_29, and NetworkStackNext be the target
built against system_current.

This matches the configuration that should be used when releasing the
platform (the platform should be released with a stable NetworkStack
target that uses stable APIs). Also NetworkStack is used as the build
target for the mainline module, which needs to use a stable API, so the
NetworkStack target should build against system_29 to be releasable.

Bug: 139269711
Test: Built, boots, WiFi working
Change-Id: I9c35c4a0159797718518fbeb050d7a4afab9949c