Use the test lib for NetworkStackIntegrationTests
Use TestNetworkStackLib instead of NetworkStackApiCurrentLib for
integration tests.
TestNetworkStackLib is closer to the actual NetworkStack; in particular
it uses the jarjared classes. The integration tests must use these same
classes to avoid conflicts with the framework versions.

Also use AndroidManifestBase instead of AndroidManifest in
TestNetworkStackLib, because the real AndroidManifest contains
privileged permissions that a test (which is not installed in
/system/priv-app) should not request; and tests should not declare
services that conflict with the real NetworkStack services.

Test: atest NetworkStackIntegrationTests FrameworksNetIntegrationTests
Change-Id: I7f4896c71bb244b81a9d232d369d6b215cd61903