Fix sll_protocol in DhcpClient packet socket
The broadcast socket address was created with a deprecated API that does
not set sll_protocol. This causes packets to be filtered incorrectly in
packet capture, typically with tcpdump.

This change only affects device with API > Q, since Q does not have the
proper API to set the protocol.

Also includes fixing the build path that was referencing apishim/current
(which does not exist) instead of apishim/30 (the correct directory).
SocketUtilsShim can be used on the system_current target with that fix.

See change: I07887b82e0e32aadb0cbb9f930f2b2fa3e277ca9

Bug: 133196453
Test: manual; on AOSP with the new build, packet capture has the right
      protocol. When installed on a Q release build, behavior is
      unchanged.

Change-Id: Iaafd0b3935473b1cfab61b49b2f107e71e01de2d