Revert "Revert "Fix sll_protocol in DhcpClient packet socket""
Fix sll_protocol in DhcpClient packet socket

The broadcast socket address was created with a deprecated API that does
not set sll_protocol. This causes packets to be filtered incorrectly in
packet capture, typically with tcpdump.

This change only affects device with API > Q, since Q does not have the
proper API to set the protocol.

Also includes fixing the build path that was referencing apishim/current
(which does not exist) instead of apishim/30 (the correct directory).
SocketUtilsShim can be used on the system_current target with that fix.

See change: I07887b82e0e32aadb0cbb9f930f2b2fa3e277ca9

Bug: 133196453
Test: manual; on AOSP with the new build, packet capture has the right
      protocol. When installed on a Q release build, behavior is
      unchanged.

This reverts commit 5ad0944d24e8dabc7fa56c3325e9200669889641.

Reason for revert:
Roll forward the previous change, as the build breakage has been
addressed. The stable build target is now used to build the
NetworkStack mainline module, which allows this change to be merged.

Change-Id: Ie933050ef269ce2159f3f44ac4db1f9aae134faf