Make IpClientIntegrationTest talk to the real netd
Currently, IpClientIntegrationTest does not talk to the real netd
because of a race on shutdown. It can happen that while setUp is
bringing up the testNetworkManager under the shell permissioni
identity, the IpClient from the previous test iss still
completing shutdown, and because the shell is not allowed to talk
to netd, IpClient was getting a SecurityException and crashing
the test.

Test: atest NetworkStackIntegrationTests
Change-Id: I8255ffec72c6e180aae58dbf23ebf39098a75c96