Pass the supportQ flag to stats-log-api-gen
Statsd is becoming a Mainline module in R. As part of that effort,
the logging schema has been updated. However, mainline modules
using statsd that are updated in Q still need to use the legacy schema.
So we've added support for that in the autogeneration tool by passing a
flag to support legacy logging schema.

Note: this change is not needed if this module will not be updated in Q.

Bug: 145624233
Test: m
Test: flashes successfully
Change-Id: Ib855001f791c28ec5166bb4400119bd04783dbb1