NetworkStackTest: Include both 32bit and 64bit lib
This is necessary to have the tests run on cf_x86_phone, even though the
test suite target is test_suites_x86_64.
Only cf_x86_phone (32-bit) can be used for test coverage at the moment,
but there is no 32-bit test suite target. This is a workaround until
test coverage can be calculated on a 64-bit target.

Test: atest NetworkStackTest

Change-Id: I2e7c5e809c795921669e828c2c0b2db302eb43a4