Update NetworkStack, NetworkStackNext SDK versions
Set NetworkStackNext version to 300000000: it effectively targets R
APIs, so as per the version scheme, the leading digits should be 30 and
not 29.
This prevents accidental updates of NetworkStackNext with NetworkStack,
as R devices should stay on NetworkStackNext to provide test coverage
for new R APIs.

Set min_sdk_version and target_sdk_version to 29 on targets that build
against stable SDK.
Remove the above attributes on targets that build against current SDK;
the build system will automatically assign the appropriate min_sdk and
target_sdk for those.

This change also changes TestNetworkStackLib to use the release API
level instead of the development API level. This makes the tests closer
to the updates that are shipped, although it prevents the integration
tests from exercising development API-only logic. This is necessary as
the tests cannot both be compatible with stable API devices, and depend
on an android_library with a development min_sdk.

Test: aapt dump badging shows the new values for NetworkStack,
      and for NetworkStackNext: NetworkStackNext has min_sdk and
      target_sdk "R", NetworkStack is 29.
      atest NetworkStackIntegrationTests FrameworksNetIntegrationTests
Change-Id: I9daed95697d21fb79af2514bbb0abc57bc6a4b5e