Include Bundle values with ConnectivityDiagnostics calls.
ConnectivityDiagnostics allows PersistableBundles to be passed from
NetworkMonitor to ConnectivityDiagnosticsCallbacks that are registered
with the System in ConnectivityService. This change populates the
bundles passed from NetworkMonitor with values relating to
ConnectivityDiagnosticsManager#ConnectivityReport and
ConnectivityDiagnosticsManager#DataStallReport using keys defined in
those classes. The keys used for the bundles are defined in
ConstantsShim.

Bug: 147249364
Test: compiles
Test: atest NetworkStackTests
Change-Id: Id007ab38a20c6a107c9528e3e861dafa191797c3
Merged-In: Id007ab38a20c6a107c9528e3e861dafa191797c3