Fix NetworkStackTests compatibility with Q
NetworkStackTests use some JNI code for APF tests, which does not link
all libc++ dependencies statically. Because of this, the test stopped
passing on Q.
Compile the JNI code against stable SDK to guarantee backwards
compatibility.

Also fix usage of the CellIdentity* constructors for Q devices: the
constructors have new parameters in later APIs. Use reflection on older
devices to reference the old constructor; the reflection call is not
likely to break as there should be no change to the Q constructor in the
future.

Bug: 148753875
Test: atest NetworkStackTests with a Q device connected
Change-Id: Ie35d6a2618847fa1de4c093c21df7c38584bbaef