Add a test runner to ignore tests on old SDKs
Generally DevSdkIgnoreRule should be used for that purpose (using rules
is preferable over replacing the test runner), however JUnit runners
inspect all methods in the test class before processing test rules. This
may cause issues if the test methods are referencing classes that do not
exist on the SDK of the device the test is run on.

Test: tests in change Iddd00e1c85abe767b1a41a1761d3266ba322dba6
Bug: 150918852
Change-Id: Ia8df394ed65fad3a3333da32c868c4623975ff79