Fix the incorrect behaviour when receiving a mismatched renew lease.
When the DHCP client is given a new IPv4 address which is not its
previous network address to be extended durig RENEWING or REBINDING
state, should roll back to INIT state and restart DHCP configuration.
That's the intended behaviour with this fix (notify provisioning failure
and then exit from receivePacket() immediately, DO NOT accept this renew
lease).

Otherwise, the new given network address will be accepted and noitfy
IpClient to configure this new address on interface, which violates
RFC2131#4.4.5. The incorect call graph will be:

- receiving mismatched renew lease:
    - setDhcpLeaseExpiry
    - acceptDhcpResults
    - notifySuccess
    - transitionTo(mDhcpBoundState)
    - keep using this new address

Bug: 131797393
Test: atest NetworkStackTests
Change-Id: I6fd37e6af1f6bf3a8ce8de7b9a1ef4731de5db27