ipmemorystore-aidl-interfaces is marked as available to platform
The lib is depended on by NetworkStackNext which is part of the platform.
With b/153073816, we will be checking the availability to the platform,
i.e. platform-available module depending on platform-unavailable module
causes a build error.

Prepare for the change by correctly marking that this interface lib is
available to the platform.

Bug: 153073816
Test: m
Change-Id: I946c9c874b381a8679071f87060d4e4739875f82