Refactor a new NetlinkMonitor class out of IpNeighborMonitor.
This CL adds a new NetlinkMonitor that is a subclass of
PacketReader and a superclass of IpNeighborMonitor. The class can
be used to implement other simple "listen to netlink broadcasts
and take action on them" classes without requiring much new code
to be written.

Bug: 153694684
Test: atest NetworkStackTests:IpReachabilityMonitorTest
Merged-In: Ia7ffb0fab4b116f4e238ccc5f5da92a04c4c86e2
Change-Id: Ia7ffb0fab4b116f4e238ccc5f5da92a04c4c86e2