Add test for NetworkStackService dumpsys version
The test verifies that each method on INetworkStackConnector updates the
version received from the remote, and that the output of dumpsys version
matches the expected template.
This logic will be changed in R to include interface hashes, so a test
is important to verify that Q behavior remains the same to avoid
breaking Q conformance tests.

Test: atest NetworkStackTests:NetworkStackServiceTest
Bug: 137328719
Change-Id: Icaab91d2cb4c62930f969612545e369f09f1b0c7