Skip tests if the shims are too old for them to pass.
Two tests in IpClientIntegrationTest depend on methods that
require R. These tests are already ignored on Q devices. However,
when running NetworkStackIntegrationTests (as opposed to
NetworkStackNextIntegrationtests), the tests fail because they
compile against the Q shims, and the code behaves the same way as
it does on Q.

Skip these tests by checking the version of the shims.

Bug: 152723363
Test: atest NetworkStackIntegrationTests has two fewer failures
Merged-In: I388bd58e79cd4310ad286bcc5688e96b16787be0
Change-Id: I388bd58e79cd4310ad286bcc5688e96b16787be0