Deflake testRaRdnss.
This test is flaky because sometimes provisioning succeeds as soon
as the link-local and stable address appear. This causes future
expectations on LinkProperties update callbacks to get confused
because they see the addition of the privacy address instead of
the change they are expecting.

Ensure that if provisioning completes with two addresses we
explicitly wait for and consume the update that contains the
privacy address.

Bug: 152723363
Test: atest NetworkStackIntegrationTests:IpClientIntegrationTest#testRestoreInitialInterfaceMtu_WithException --iterations 100
Original-Change: http://aosp/1295487
Merged-In: Ibdd16ce0285db20bd700bf4e78762b922b4661d4
Change-Id: Ibdd16ce0285db20bd700bf4e78762b922b4661d4