Do not stop FdEventsReader on handlePacket error
Previously if handlePacket threw an exception, FdEventsReader would
stop reading and close the file descriptor.
This is not helpful as users of this class may encounter errors
processing a particular packet (typically), but they generally should
not stop completely for that reason.

Also address comments from previous change (Exception instead of
Throwable: Throwable should not be caught in general; event -> events).

Bug: 152842850
Test: atest NetworkStackTests
Change-Id: I14abe8163ecdfd5251731915f10ea374bfa38b70