Wait for IpClient.StoppedState#enter before tests
IpClientIntegrationTest is flaky because some mocks that IpClient uses,
like mNetd, can be stubbed while IpClient is initializing in its handler
thread. This generates UnfinishedStubbingExceptions as IpClient is
calling methods on the mock while the test is trying to stub additional
methods.

Wait for IpClient to become idle before doing any further setup.

Bug: 152723363
Test: atest NetworkStackCoverageTests:IpClientIntegrationTests
          --rerun-until-failure [lots] on Q Cuttlefish
Change-Id: I90206a16fff1fa12ca7e37e36d49163b6a83e5c1