Don't crash if the PREF64 expires when IpClient is stopped.
Currently the code does not cancel the PREF64 alarm when IpClient
is stopped. If the alarm fires when IpClient has disconnected, it
will call updatePref64 with a null prefix, which will crash.

Ensure that the alarm is cancelled when IpClient is stopped.

Bug: 153694684
Bug: 155559286
Test: new unit tests
Change-Id: I397eb36c1a8ecf1324de33e720577fe2e4e07056
Merged-In: I397eb36c1a8ecf1324de33e720577fe2e4e07056