Avoid rely on NETWORK_STACK permission for InetDiagSocketTest
ConnectivityManager#getConnectionOwnerUid required NETWORK_STACK
permission. Move related tests from InetDiagSocketTest(unit test) to
InetDiagSocketIntegrationTest. The test is an integration test and
not a unit test, as its purpose is to exercise the getConnectionOwnerUid
platform API. Since network stack unit test is shared with
TetheringCoverageTests, this is part of work to avoid rely on platform
cert for TetheringCoverageTests.

Bug: 156866746
Test: TetheringCoverageTest, NetworkStackNextTests,
      NetworkStackCoverageTests
Change-Id: Ic38852839ead951972daa24f27ac95376787a72c