Do not revalidate before network is ready
ConnectivityService sends notifyLinkPropertiesChanged before
notifyNetworkConnected. When a captive portal URL is present, this
causes NetworkMonitor to revalidate, even though the network is not
ready to start validating (DNS servers have not been set yet in
particular). ConnectivityService does this because the sending new
LinkProperties to NetworkMonitor is part of the standard LinkProperties
update flow; NetworkMonitor should be resilient to such behavior.

Test: atest NetworkMonitorTest, manual (flashed, wifi working)
Bug: 156697983
Change-Id: I0619d37a3374726b77e162d174c5b12659db3bbb