Use the latest stable netd AIDL interface in NetworkStack.
This is not needed by NetworkStack itself, but is needed by
Tethering. Update the NetworkStack version as well in order to
avoid multiple different versions of the AIDL generated classes
in the same process. While this should be safe, because the two
APKs have two different ClassLoaders, it still seems better to
be consistent.

AOSP and master are already using -unstable instead of -V3.
In order not to change master, the Merged-In tag is set to the CL
that switched AOSP (and master) to -unstable.

Test: m
Bug: 150736748
Merged-In: I70e22e6d78feac8f153e746fad5bb512fc31bde1
Change-Id: I5ccafd727bc53c0c467842e51924a6e755880317